Fixed free space number by changing a int to a float.
git-svn-id: https://burn-osx.svn.sourceforge.net/svnroot/burn-osx/trunk@46 df5e183e-a46d-47f3-acf2-86dc86df8d71

+++ burn/Source/KWDocument.m
@@ -462,7 +462,7 @@ return [NSArray arrayWithObjects:NSToolbarFlexibleSpaceItemIdentifier,@"Main",NS
 {
 	if (device)
 	{
-	int space;
+	float space;
 	
 		if ([[[device status] objectForKey:DRDeviceMediaStateKey] isEqualTo:DRDeviceMediaStateMediaPresent])
 		{
